moved generic code to pipelines package
