+++ launcher/config_test.go
@@ -314,7 +314,6 @@ func TestDefaultConfig(t *testing.T) {
 		MetricExporterEndpointInsecure: false,
 		MetricReportingPeriod:          "30s",
 		MetricsEnabled:                 true,
-		AccessToken:                    "",
 		LogLevel:                       "info",
 		Propagators:                    []string{"b3"},
 		Resource:                       resource.NewWithAttributes(semconv.SchemaURL, attributes...),
@@ -350,7 +349,6 @@ func TestEnvironmentVariables(t *testing.T) {
 		MetricExporterEndpoint:         "metrics-url",
 		MetricExporterEndpointInsecure: true,
 		MetricReportingPeriod:          "30s",
-		AccessToken:                    "token",
 		LogLevel:                       "debug",
 		Propagators:                    []string{"b3", "w3c"},
 		Resource:                       resource.NewWithAttributes(semconv.SchemaURL, attributes...),
@@ -398,7 +396,6 @@ func TestConfigurationOverrides(t *testing.T) {
 		MetricExporterEndpointInsecure: false,
 		MetricReportingPeriod:          "30s",
 		Headers:                        map[string]string{"lightstep-access-token": "override-access-token"},
-		AccessToken:                    "override-access-token",
 		LogLevel:                       "info",
 		Propagators:                    []string{"b3"},
 		Resource:                       resource.NewWithAttributes(semconv.SchemaURL, attributes...),
