+++ tools/go.mod
@@ -1,6 +1,6 @@
 module github.com/lightstep/opentelemetry-go/tools
 
-go 1.14
+go 1.16
 
 require (
 	github.com/client9/misspell v0.3.4
